requirements.txt
git+https://github.com/Avantis-Labs/avantis_trader_sdk.git
asyncio
websockets>=11.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
click>=8.0.0
aiofiles>=23.0.0
pandas>=2.0.0
numpy>=1.24.0
ta-lib>=0.4.0
ccxt>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
